moved write_cube to shared.io
this fixes a circular import of molbe.BE
molbe.BE depends on shared/helper, so we cannot import it in
shared/helper.py
